Change .env.production hostnames
I had been using ${COMPOSE_PROJECT_NAME}-db-1 as the DB (and similar for redis and es)  in order to match my long dns names for my system, but when used with a dedicated internal network, the short names suffice and are the same regardless of docker or compose versions. I have one system that the long name uses underscores "_" and another that uses hyphens "-" to build up the long names, using the short names works if using a dedicated internal network.